requirements.txt
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mutagen==1.47.0
pycryptodomex==3.20.0
pyee==11.1.0
python-ffmpeg==2.0.10
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.3.10
